refactor: WinnerJudgeImpl 패키지 이동
