+++ src/main/java/racingcar/domain/race/WinnerJudgeImpl.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.race;
 
 import java.util.List;
 import java.util.stream.Collectors;
