+++ src/test/java/racingcar/domain/WinnerJudgeTest.java
@@ -6,7 +6,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.WinnerJudgeImpl;
+import racingcar.domain.race.WinnerJudgeImpl;
 import racingcar.domain.car.Car;
 
 class WinnerJudgeTest {
